reduce_resolution/requirements.txt
flask==2.0.1
Pillow==9.0.0
PyMuPDF==1.19.0